pyproject.toml
[project]
name = "page-editor-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "opperai>=0.31.2",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
]
